Consegna M5

+++ AMM/web/WEB-INF/db_definition/crea_db.sql
@@ -3,7 +3,7 @@ CREATE TABLE Utenti(
 	password VARCHAR(32) NOT NULL,
 	saldo DOUBLE,
 	tipo CHAR(1) NOT NULL,
-	feedback INTEGER
+	feedback DOUBLE
 );
 
 CREATE TABLE Articoli(
